# Fix performance for gatheInfo func, move regexp to global var for onc…