…e regexp compilation.

Benchmark result:

goos: linux
goarch: amd64
pkg: github.com/kelseyhightower/envconfig

benchmark                 old ns/op     new ns/op     delta
BenchmarkGatherInfo-4     756307        535401        -29.21%

benchmark                 old allocs     new allocs     delta
BenchmarkGatherInfo-4     349            251            -28.08%

benchmark                 old bytes     new bytes     delta
BenchmarkGatherInfo-4     64467         15425         -76.07%